Extract recovery related endpoints from upgrade endpoint.
Move to SdxRecoveryController.
