+++ datalake-api/src/main/java/com/sequenceiq/sdx/api/endpoint/SdxUpgradeEndpoint.java
@@ -2,7 +2,6 @@ package com.sequenceiq.sdx.api.endpoint;
 
 import javax.validation.Valid;
 import javax.ws.rs.Consumes;
-import javax.ws.rs.GET;
 import javax.ws.rs.POST;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
@@ -12,9 +11,6 @@ import javax.ws.rs.core.MediaType;
 import org.springframework.validation.annotation.Validated;
 
 import com.sequenceiq.cloudbreak.jerseyclient.RetryAndMetrics;
-import com.sequenceiq.sdx.api.model.SdxRecoverableResponse;
-import com.sequenceiq.sdx.api.model.SdxRecoveryRequest;
-import com.sequenceiq.sdx.api.model.SdxRecoveryResponse;
 import com.sequenceiq.sdx.api.model.SdxUpgradeRequest;
 import com.sequenceiq.sdx.api.model.SdxUpgradeResponse;
 
@@ -40,30 +36,5 @@ public interface SdxUpgradeEndpoint {
     @ApiOperation(value = "upgrades the data lake cluster", nickname = "upgradeDatalakeClusterByCrn")
     SdxUpgradeResponse upgradeClusterByCrn(@PathParam("crn") String crn,
             @Valid SdxUpgradeRequest upgradeSdxClusterRequest);
-
-    @POST
-    @Path("{name}/recover")
-    @Produces(MediaType.APPLICATION_JSON)
-    @ApiOperation(value = "recovers the data lake upgrade", nickname = "recoverDatalakeCluster")
-    SdxRecoveryResponse recoverClusterByName(@PathParam("name") String name, @Valid SdxRecoveryRequest recoverSdxClusterRequest);
-
-    @POST
-    @Path("/crn/{crn}/recover")
-    @Produces(MediaType.APPLICATION_JSON)
-    @ApiOperation(value = "recovers the data lake upgrade", nickname = "recoverDatalakeClusterByCrn")
-    SdxRecoveryResponse recoverClusterByCrn(@PathParam("crn") String crn,
-            @Valid SdxRecoveryRequest recoverSdxClusterRequest);
-
-    @GET
-    @Path("{name}/recoverable")
-    @Produces(MediaType.APPLICATION_JSON)
-    @ApiOperation(value = "validates if the data lake is recoverable or not", nickname = "getClusterRecoverableByName")
-    SdxRecoverableResponse getClusterRecoverableByName(@PathParam("name") String name);
-
-    @GET
-    @Path("/crn/{crn}/recoverable")
-    @Produces(MediaType.APPLICATION_JSON)
-    @ApiOperation(value = "validates if the data lake is recoverable or not", nickname = "getClusterRecoverableByCrn")
-    SdxRecoverableResponse getClusterRecoverableByCrn(@PathParam("crn") String crn);
 
 }
